Content/getting started with conversation js sdk tweaks (#14)
* Change UUIDs

* Change code (see commit details)

- Remove username - It wasn’t being used anyway, most implementations
would use the session. Plus it was confusing that the name that came
back wasn’t the username the user entered but rather the one setup with
the API.

- Simplify HTML - Previously there was a mix of elements that shared
the same names, some as class names some as IDs. Was confusing.

- Reference but remove the call to `app.getConversations` in favour of
simpler singular `getConversation` given that we know the ID anyway.

- Replace confusing series of inline anonymous for ConversationClient
promises in favour of named functions.

- Adds datetime of message.

- More descriptive variable names.

- Remove need to bind `conversation` to global scope, also removes the
need to setup an unassigned variable at the start.

- Modernise JavaScript. Tested working in Chrome / Firefox without
transpiling.

* Further copy tweaks

* updating quickstarts with the new class approach

* nits

* addressing Adam's comments